services/lucius_dispatcher/pyproject.toml
[project]
name = "lucius-reconciliation"
version = "0.1.0"
description = "LUCIUS reconciliation service"
requires-python = ">=3.11"
dependencies = [
  "azure-servicebus>=7.11.0",
  "azure-data-tables>=12.4.0",
]

[project.scripts]
lucius-reconciliation = "app.main:main"
